Ajustes de login: App.jsx solo enrutamiento, se crea archivo pages/PasswordRecovery.jsx para ajustar errores.
